Оформили файлы в виде CommonJS
